\name{FormatString}
\alias{FormatString}
\encoding{UTF-8}
\title{FormatString}
\description{
  Formats character strings so they can be safely used in SQL queries.
}

\usage{
FormatString(con, s)
}
\arguments{
  \item{con}{Connection object to the database.}
  \item{s}{The character string to format.}
}
\value{
  The character string formatted.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{format}
\keyword{sql}
\keyword{string}
\keyword{character}